TST: rename pytest_plugins -> pytest_config
The name `pytest_plugins` is an internally used reserved name that
our package messes up, see https://github.com/pytest-dev/pytest/issues/3899
